real bitrate progress
Put the real bitrate in the progress file.  In case of VBR, jack used to
put the bitrate of CBR into the jack.progress file.  (Actually, why does
track[RATE] not always contain the real rate?)  Or maybe we should put the
chosen quality rating in the progress file in case of OGG?

Debian patch: 31_real_bitrate_progress.patch