fix datatrack rename
Fixes that "jack -R -t x" fails when x is a data track and CD not in drive.
jack doesn't realize that it is in fact a data track.  We can fix this by
moving the code around a bit so the jack.progress file is parsed first,
then jack will know it's not a valid audio track.  Debian #341889

Depends on 70_fix_upd_progress.patch.

Debian patch: 75_fix_datatrack_rename.patch