dir template yg
Make %y and %g actually work dir_template (when the user doesn't explicitly
specify it with -G and -Y).   First we need to set it when doing the FreeDB
lookup, and secondly we need to use the right variable for the replacement.
_id3_genre_txt is not set anywhere at all.  Debian #351574

Depends on 76_mkdirname_cleanup.patch and 84_dyear_dgenre.patch

Debian patch: 85_dir_template_yg.patch