correct-category-list
Cédric Boutillier <cedric.boutillier@gmail.com> reports:

When I try to send a submission to a freedb server, the list of
categories appearing on the screen does not contain 'blues' (which is a
freedb category) and contains 'null' (which is not a freedb category).

I think that a few changes must be made to jack_freedb.py:
* when the list is fetched from the server, it is unnecessarily
  initialised with "null"
* subtract 1 to the index to get the correct category
* When the list is displayed, the index is running from 1 to len(cat)-1,
  missing the first item.

The attached patch solved the problem for me, to be applied after the
already very long list of patches.

See http://bugs.debian.org/536165

Debian patch: 99a_correct-category-list.patch